Dual Mesh now uses an array of DelaunayTriangle structs for its triangles array.
If you still want the old indices, they are stored as AIndex, BIndex, and CIndex inside each index of the DelaunayTriangle struct.

+++ Source/PolygonalMapGenerator/Public/IslandMap.h
@@ -106,7 +106,7 @@ public:
 	UPROPERTY()
 	TArray<bool> r_ocean;
 	UPROPERTY()
-	TArray<float> t_coastdistance;
+	TArray<int32> t_coastdistance;
 	UPROPERTY()
 	TArray<float> t_elevation;
 	UPROPERTY()
